imp: add history() under TestHost
