+++ ibc-testkit/src/hosts/mod.rs
@@ -1,6 +1,7 @@
 pub mod mock;
 pub mod tendermint;
 
+use alloc::collections::VecDeque;
 use core::fmt::Debug;
 use core::ops::Add;
 use core::time::Duration;
@@ -39,7 +40,7 @@ pub type HostConsensusState<H> = <HostHeader<H> as TestHeader>::ConsensusState;
 /// TestHost is a trait that defines the interface for a host blockchain.
 pub trait TestHost: Debug + Sized {
     /// The type of block produced by the host.
-    type Block: TestBlock + Debug;
+    type Block: TestBlock;
 
     /// The type of client state produced by the host.
     type ClientState: Into<AnyClientState> + Debug;
@@ -53,21 +54,36 @@ pub trait TestHost: Debug + Sized {
     /// Build a new host with the given parameters.
     fn build(params: HostParams) -> Self;
 
+    fn history(&self) -> &VecDeque<Self::Block>;
+
     /// The chain identifier of the host.
     fn chain_id(&self) -> &ChainId;
 
     fn block_time(&self) -> Duration;
 
-    fn is_empty(&self) -> bool;
     fn genesis_timestamp(&self) -> Timestamp;
-    fn latest_block(&self) -> Self::Block;
-    fn get_block(&self, target_height: &Height) -> Option<Self::Block>;
-    fn push_block(&mut self, block: Self::Block);
-    fn prune_block_till(&mut self, height: &Height);
+
+    fn is_empty(&self) -> bool {
+        self.history().is_empty()
+    }
 
     fn latest_height(&self) -> Height {
         self.latest_block().height()
     }
+
+    fn latest_block(&self) -> Self::Block {
+        self.history().back().cloned().expect("no error")
+    }
+
+    fn get_block(&self, target_height: &Height) -> Option<Self::Block> {
+        self.history()
+            .get(target_height.revision_height() as usize - 1)
+            .cloned() // indexed from 1
+    }
+
+    fn push_block(&mut self, block: Self::Block);
+
+    fn prune_block_till(&mut self, height: &Height);
 
     /// Triggers the advancing of the host chain, by extending the history of blocks (or headers).
     fn advance_block(&mut self, commitment_root: Vec<u8>, params: &Self::BlockParams) {
@@ -92,20 +108,6 @@ pub trait TestHost: Debug + Sized {
         self.push_block(new_block);
     }
 
-    fn validate(&self) -> Result<(), String> {
-        // Check that headers in the history are in sequential order.
-        let latest_height = self.latest_height();
-        let mut current_height = Height::min(latest_height.revision_number());
-
-        while current_height <= latest_height {
-            if current_height != self.get_block(&current_height).expect("no error").height() {
-                return Err("block height does not match".to_owned());
-            }
-            current_height = current_height.increment();
-        }
-        Ok(())
-    }
-
     /// Generate a block at the given height and timestamp, using the provided parameters.
     fn generate_block(
         &self,
@@ -121,6 +123,20 @@ pub trait TestHost: Debug + Sized {
         latest_height: &Height,
         params: &Self::LightClientParams,
     ) -> Self::ClientState;
+
+    fn validate(&self) -> Result<(), String> {
+        // Check that headers in the history are in sequential order.
+        let latest_height = self.latest_height();
+        let mut current_height = Height::min(latest_height.revision_number());
+
+        while current_height <= latest_height {
+            if current_height != self.get_block(&current_height).expect("no error").height() {
+                return Err("block height does not match".to_owned());
+            }
+            current_height = current_height.increment();
+        }
+        Ok(())
+    }
 }
 
 /// TestBlock is a trait that defines the interface for a block produced by a host blockchain.
