+++ ibc-testkit/src/hosts/tendermint.rs
@@ -36,9 +36,9 @@ pub struct TendermintHost {
 
 impl TestHost for TendermintHost {
     type Block = TendermintBlock;
+    type ClientState = ClientState;
     type BlockParams = BlockParams;
     type LightClientParams = MockClientConfig;
-    type ClientState = ClientState;
 
     fn build(params: HostParams) -> Self {
         let HostParams {
@@ -55,6 +55,10 @@ impl TestHost for TendermintHost {
         }
     }
 
+    fn history(&self) -> &VecDeque<Self::Block> {
+        &self.history
+    }
+
     fn chain_id(&self) -> &ChainId {
         &self.chain_id
     }
@@ -63,22 +67,8 @@ impl TestHost for TendermintHost {
         self.block_time
     }
 
-    fn is_empty(&self) -> bool {
-        self.history.is_empty()
-    }
-
     fn genesis_timestamp(&self) -> Timestamp {
         self.genesis_timestamp
-    }
-
-    fn latest_block(&self) -> Self::Block {
-        self.history.back().cloned().expect("Never fails")
-    }
-
-    fn get_block(&self, target_height: &Height) -> Option<Self::Block> {
-        self.history
-            .get(target_height.revision_height() as usize - 1)
-            .cloned() // indexed from 1
     }
 
     fn push_block(&mut self, block: Self::Block) {
